chore: migrate to cmake
