+++ .gitignore
@@ -1,27 +1,35 @@
-*.la
-*.lo
-*.o
-*_test
-*~
-
-.DS_Store
-
-.deps
-.libs
-
-Doxyfile
+CMakeLists.txt.user
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
 Makefile
-Makefile.in
-aclocal.m4
-api-docs
-autom4te.cache
-build
-config.h
-config.h.in
-config.log
-config.status
-configure
-libtool
-lutok.pc
-releases
-stamp-h1
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+_deps
+CMakeUserPresets.json
+
+# User-specific stuff
+.idea/**/workspace.xml
+.idea/**/tasks.xml
+.idea/**/usage.statistics.xml
+.idea/**/dictionaries
+.idea/**/shelf
+
+# Generated files
+.idea/**/contentModel.xml
+
+# Sensitive or high-churn files
+.idea/**/dataSources/
+.idea/**/dataSources.ids
+.idea/**/dataSources.local.xml
+.idea/**/sqlDataSources.xml
+.idea/**/dynamic.xml
+.idea/**/uiDesigner.xml
+.idea/**/dbnavigator.xml
+
+# CMake
+cmake-build-*/
+build/
