+++ Doxyfile
@@ -35,16 +35,16 @@ EXTRACT_STATIC = YES
 EXPAND_ONLY_PREDEF = YES
 FILE_PATTERNS = *.cpp *.hpp *.ipp
 GENERATE_LATEX = NO
-GENERATE_TAGFILE = @top_builddir@/api-docs/api-docs.tag
-INPUT = @top_srcdir@
+GENERATE_TAGFILE = build/api-docs/api-docs.tag
+INPUT = .
 INPUT_ENCODING = ISO-8859-1
 JAVADOC_AUTOBRIEF = YES
 MACRO_EXPANSION = YES
-OUTPUT_DIRECTORY = @top_builddir@/api-docs
+OUTPUT_DIRECTORY = build/api-docs
 OUTPUT_LANGUAGE = English
 PREDEFINED = "UTILS_UNUSED_PARAM(name)=unused_ ## name"
-PROJECT_NAME = "@PACKAGE_NAME@"
-PROJECT_NUMBER = @VERSION@
+PROJECT_NAME = "Lutok"
+PROJECT_NUMBER = 0.6.3
 QUIET = YES
 RECURSIVE = NO
 SHORT_NAMES = YES  # Cope with gnutar limitations during 'make dist'.
