--- lutok.pc.in
@@ -1,8 +1,0 @@
-includedir=__INCLUDEDIR__
-libdir=__LIBDIR__
-
-Name: lutok
-Description: Lightweight C++ API for Lua
-Version: __VERSION__
-Cflags: __LUA_CFLAGS__ -I${includedir}
-Libs: __LUA_LIBS__ -L${libdir} -llutok
