+++ scripts/ci-build/01_configure.sh
@@ -4,21 +4,24 @@
 #
 # SPDX-License-Identifier: BSD-2-Clause
 
-# Step 1: run autoconf/configure so the remaining of the build/test steps can
+# Step 1: run cmake configuration so the remaining build/test steps can
 # be completed.
 #
-# Splitting off this step allows any configure issues to be found quickly and
-# triaged more effectively.
+# Splitting off this step allows any configuration issues to be found quickly
+# and triaged more effectively.
 
-autoreconf_args="-isv"
-if [ -d /usr/local/share/aclocal ]; then
-    autoreconf_args="${autoreconf_args} -I/usr/local/share/aclocal"
+set -eux
+
+CMAKE_ARGS="-DBUILD_TESTING=ON -DBUILD_DOCS=OFF"
+
+# Developer mode is auto-enabled for git repos, but we can force it
+if [ -n "${DEVELOPER_MODE:-}" ]; then
+    CMAKE_ARGS="${CMAKE_ARGS} -DDEVELOPER_MODE=${DEVELOPER_MODE}"
 fi
-# shellcheck disable=SC2086
-autoreconf ${autoreconf_args}
 
-if ! ./configure --enable-atf; then
-    cat config.log || true
+if ! cmake -B build -S . ${CMAKE_ARGS}; then
+    cat build/CMakeFiles/CMakeError.log || true
+    cat build/CMakeFiles/CMakeOutput.log || true
     exit 1
 fi
