+++ tests/c_gate_test.cpp
@@ -26,13 +26,13 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "c_gate.hpp"
+#include <lutok/c_gate.hpp>
 
 #include <atf-c++.hpp>
 #include <lua.hpp>
 
-#include "state.ipp"
-#include "test_utils.hpp"
+#include <lutok/state.ipp>
+#include <lutok/test_utils.hpp>
 
 
 ATF_TEST_CASE_WITHOUT_HEAD(connect);
