+++ tests/operations_test.cpp
@@ -26,15 +26,15 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "operations.hpp"
+#include <lutok/operations.hpp>
 
 #include <fstream>
 
 #include <atf-c++.hpp>
 
-#include "exceptions.hpp"
-#include "state.ipp"
-#include "test_utils.hpp"
+#include <lutok/exceptions.hpp>
+#include <lutok/state.ipp>
+#include <lutok/test_utils.hpp>
 
 
 namespace {
