+++ tests/stack_cleaner_test.cpp
@@ -26,7 +26,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "stack_cleaner.hpp"
+#include <lutok/stack_cleaner.hpp>
 
 #include <atf-c++.hpp>
